rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.7.0
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
numpy==1.24.3
pydantic==2.5.0
huggingface-hub==0.23.0
faiss-cpu==1.7.4
